refactor: remove delete functionality for Jira issues and Confluence pages
This commit removes the `delete_issue` method from the `IssuesMixin` class and the `delete_page` method from the `PagesMixin` class, along with their associated server tools. The changes also include the removal of related tests and references in the README.md.

This commit aims to remove tools that are not used in our contexts and also remove tools that present some sort of inherent security issue.